+++ CopilotKit/packages/react-ui/src/components/dev-console/console.tsx
@@ -1,6 +1,6 @@
 "use client";
 
-import { useCopilotContext } from "@copilotkit/react-core";
+import { useCopilotContext, useCopilotMessagesContext } from "@copilotkit/react-core";
 import {
   getPublishedCopilotKitVersion,
   logActions,
@@ -227,9 +227,10 @@ export default function DebugMenuButton({
   mode: "full" | "compact";
 }) {
   const context = useCopilotContext();
+  const messagesContext = useCopilotMessagesContext();
 
   return (
-    <div className="bg-black fixed top-24 w-52 text-right">
+    <div className="bg-black top-24 w-52 text-right">
       <Menu>
         <MenuButton className={`copilotKitDebugMenuButton ${mode === "compact" ? "compact" : ""}`}>
           {mode == "compact" ? "Debug" : <>Debug {ChevronDownIcon}</>}
@@ -252,7 +253,10 @@ export default function DebugMenuButton({
             </button>
           </MenuItem>
           <MenuItem>
-            <button className="copilotKitDebugMenuItem" onClick={() => logMessages(context)}>
+            <button
+              className="copilotKitDebugMenuItem"
+              onClick={() => logMessages(messagesContext)}
+            >
               Log Messages
             </button>
           </MenuItem>