+++ examples/coagents-qa/ui/app/page.tsx
@@ -7,7 +7,7 @@ import "@copilotkit/react-ui/styles.css";
 export default function Home() {
   return (
     <main className="flex flex-col items-center justify-between">
-      <CopilotKit runtimeUrl="/api/copilotkit">
+      <CopilotKit runtimeUrl="/api/copilotkit" agent="email_agent">
         <Mailer />
       </CopilotKit>
     </main>