+++ CopilotKit/packages/react-textarea/CHANGELOG.md
@@ -1,5 +1,47 @@
 # ui
 
+## 1.3.6
+
+### Patch Changes
+
+- 1. Removes the usage of the `crypto` Node pacakge, instaed uses `uuid`. This ensures that non-Next.js React apps can use CopilotKit.
+  2. Fixes Nest.js runtime docs
+
+- Updated dependencies
+  - @copilotkit/react-core@1.3.6
+  - @copilotkit/runtime-client-gql@1.3.6
+  - @copilotkit/shared@1.3.6
+
+## 1.3.5
+
+### Patch Changes
+
+- Improve CoAgent state render
+- Updated dependencies
+  - @copilotkit/react-core@1.3.5
+  - @copilotkit/runtime-client-gql@1.3.5
+  - @copilotkit/shared@1.3.5
+
+## 1.3.4
+
+### Patch Changes
+
+- Add followUp property to useCopilotAction
+- Updated dependencies
+  - @copilotkit/react-core@1.3.4
+  - @copilotkit/runtime-client-gql@1.3.4
+  - @copilotkit/shared@1.3.4
+
+## 1.3.3
+
+### Patch Changes
+
+- Impvovements to error handling and CoAgent protocol
+- Updated dependencies
+  - @copilotkit/react-core@1.3.3
+  - @copilotkit/runtime-client-gql@1.3.3
+  - @copilotkit/shared@1.3.3
+
 ## 1.3.2
 
 ### Patch Changes